Wtf is going on

+++ EpilepsyApp/DTO/PythonEcgProcessedMeasurements.cs
@@ -7,7 +7,7 @@ public class PythonEcgProcessedMeasurements
 	public float CSI50 { get; set; }
 	public float CSI100 { get; set; }
 	public float ModCSI100 { get; set; }
-	public List<double> ProcessedEcgChannel1 { get; set; }
-	public List<double> ProcessedEcgChannel2 { get; set; }
-	public List<double> ProcessedEcgChannel3 { get; set; }
+	public List<int> ProcessedEcgChannel1 { get; set; }
+	public List<int> ProcessedEcgChannel2 { get; set; }
+	public List<int> ProcessedEcgChannel3 { get; set; }
 }
